% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tauSurface.R
\name{bootstrapCoefs}
\alias{bootstrapCoefs}
\title{Show estimated model parameters from all bootstrap iterations}
\usage{
bootstrapCoefs(tauSurface)
}
\arguments{
\item{tauSurface}{A \code{tauSurface} object returned by 
\code{\link{getTauSurface}}.}
}
\value{
matrix of parameter estimates from the bootstrap iterations as 
returned by \code{\link{getBootTaus}}
}
\description{
Show estimated model parameters from all bootstrap iterations
}
\author{
Maxim Nazarov
}
