% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.tauSurface.R
\name{contour.tauSurface}
\alias{contour.tauSurface}
\title{Contour-plot method for "tauSurface" objects}
\usage{
\method{contour}{tauSurface}(x, digits = 4, ...)
}
\arguments{
\item{x}{A \code{tauSurface} object returned by \code{\link{getTauSurface}}.}

\item{digits}{Number of digits used in \code{\link[base]{format}} for the 
dose labels.}

\item{...}{Further parameters, currently not used.}
}
\value{
a ggplot2 object
}
\description{
Contour-plot method for "tauSurface" objects
}
\details{
The function returns a heatmap-like plot displaying the interaction 
index (tau) estimates and pointwise confidence intervals numerically. 
Blue and red colours are used to indicate areas of synergy (if confidence 
interval lies entirely below 1) or antagonism (if confidence interval lies 
entirely above 1). Note that color intensity is determined by the absolute 
values of the interaction indices.
}
\author{
Maxim Nazarov
}
