% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.HarbronFit.R
\name{fitPlot3d}
\alias{fitPlot3d}
\title{Plot 3d surface of observations and model fit}
\usage{
fitPlot3d(fit, logScale = TRUE, useFineGrid = TRUE, showMesh = TRUE,
  widget = FALSE)
}
\arguments{
\item{fit}{A \code{HarbronFit} object returned by 
\code{\link{fitModel}}.}

\item{logScale}{Whether to use log-scale for x and y axes.}

\item{useFineGrid}{Whether to use fine grid for plotting the fitted surface 
(default), or calculate predictions only at the observed data points.}

\item{showMesh}{Whether to show 'mesh' on the plot. Note: currently doesn't 
play nicely with the \code{useFineGrid} argument, so the latter is ignored.}

\item{widget}{Whether to return a "htmlwidget" object instead of plotting on 
a 3d device.}
}
\value{
A 3d plot is shown or an object of class "htmlwidget" as returned by 
\code{\link[rgl]{rglwidget}}.
}
\description{
Plot 3d surface of observations and model fit
}
\author{
Maxim Nazarov
}
