% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{invL4}
\alias{invL4}
\title{Inverse 4-parameter log-logistic (response-dose) function}
\usage{
invL4(y, h, b, m, logEC50)
}
\arguments{
\item{y}{Response.}

\item{h}{Hill's coefficient (slope of the curve).}

\item{b}{Baseline effect (at zero dose).}

\item{m}{Maximal/asymptote effect (at infinite dose).}

\item{logEC50}{Point of inflection (in logarithmic terms).}
}
\value{
Dose level.
}
\description{
Inverse 4-parameter log-logistic (response-dose) function
}
