% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tauSurface.R
\name{print.tauSurface}
\alias{print.tauSurface}
\title{Print method for "tauSurface" objects}
\usage{
\method{print}{tauSurface}(x, ...)
}
\arguments{
\item{x}{Output of \code{\link{getTauSurface}}.}

\item{...}{Further arguments, currently not used.}
}
\description{
Print method for "tauSurface" objects
}
