% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rayData}
\alias{rayData}
\title{Example ray design drug combination data}
\format{
A data frame with 378 observations of 4 variables:
\itemize{
 \item d1: dose of the first compound
 \item d2: dose of the second compound
 \item effect: observed effect (radioactivity level)
 \item ray: a character vector with values 0, 0.2, 0.35, 0.5, 
0.65, 0.8 and 1, corresponding to the mixture factors. Each ray has 9 dose 
combinations with 6 replicates.
}
}
\description{
A dataset containing data from an in-vitro drug combination experiment in 
oncology using a ray design. The variables are as follows:
}
