% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tauSurface.R
\name{unique.tauSurface}
\alias{unique.tauSurface}
\title{Unique method for "tauSurface" objects}
\usage{
\method{unique}{tauSurface}(x, ...)
}
\arguments{
\item{x}{Output of \code{\link{getTauSurface}}.}

\item{...}{Further arguments, currently not used.}
}
\description{
"unique" method to extract unique interaction index estimates from a 
"tauSurface" object in a tabular format.
}
